+++ packages/react/Text/src/index.ts
@@ -1,0 +1,5 @@
+import { Text } from "./Text";
+
+export { Text } from "./Text";
+export type { TextProps } from "./text.types";
+export default Text;