+++ packages/styles/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./themes/base.css";
-export * from "./themes/dark.css";
 export * from "./packages";